Partition flags enum fixtures
Move the flags enum accumulator compiler fixtures beside their sole owning test class without changing the fixture declarations or method body.
